+++ imports/startup/client/index.js
@@ -2,6 +2,6 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Meteor } from 'meteor/meteor';
 
-import { App } from '/imports/ui/layouts';
+import { App } from '/imports/ui/App';
 
 Meteor.startup(() => render(<App />, document.getElementById('react-root')));
