--- imports/ui/components/Authenticated/index.js
@@ -1,1 +1,0 @@
-export { default as Authenticated } from './Authenticated';
