--- imports/ui/components/index.js
@@ -1,3 +1,0 @@
-import { Authenticated } from './Authenticated';
-
-export { Authenticated };
