--- imports/ui/layouts/App/components/header/index.js
@@ -1,1 +1,0 @@
-export { default as Header } from './Header';
