--- imports/ui/layouts/App/components/index.js
@@ -1,3 +1,0 @@
-import { Header } from './header';
-
-export { Header };
