--- imports/ui/layouts/App/index.js
@@ -1,1 +1,0 @@
-export { default as App } from './App';
